+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/entity/npc/BWGVillagerTypes.java
@@ -1,11 +1,14 @@
 package net.potionstudios.biomeswevegone.world.entity.npc;
 
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.entity.npc.VillagerType;
+import net.minecraft.world.level.biome.Biome;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
 import net.potionstudios.biomeswevegone.PlatformHandler;
 import net.potionstudios.biomeswevegone.world.level.levelgen.biome.BWGBiomes;
 
+import java.util.function.BiConsumer;
 import java.util.function.Supplier;
 
 public class BWGVillagerTypes {
@@ -13,12 +16,12 @@ public class BWGVillagerTypes {
 	public static final Supplier<VillagerType> SKYRIS = register("skyris");
 	public static final Supplier<VillagerType> SALEM = register("salem");
 
-	public static void setVillagerBWGBiomes() {
-		VillagerType.BY_BIOME.put(BWGBiomes.MOJAVE_DESERT, VillagerType.DESERT);
-		VillagerType.BY_BIOME.put(BWGBiomes.WINDSWEPT_DESERT, VillagerType.DESERT);
-		VillagerType.BY_BIOME.put(BWGBiomes.TROPICAL_RAINFOREST, VillagerType.JUNGLE);
-		VillagerType.BY_BIOME.put(BWGBiomes.SKYRIS_VALE, SKYRIS.get());
-		VillagerType.BY_BIOME.put(BWGBiomes.WEEPING_WITCH_FOREST, SALEM.get());
+	public static void setVillagerBiomes(BiConsumer<ResourceKey<Biome>, VillagerType> consumer) {
+		consumer.accept(BWGBiomes.MOJAVE_DESERT, VillagerType.DESERT);
+		consumer.accept(BWGBiomes.WINDSWEPT_DESERT, VillagerType.DESERT);
+		consumer.accept(BWGBiomes.TROPICAL_RAINFOREST, VillagerType.JUNGLE);
+		consumer.accept(BWGBiomes.SKYRIS_VALE, SKYRIS.get());
+		consumer.accept(BWGBiomes.WEEPING_WITCH_FOREST, SALEM.get());
 	}
 
 	private static Supplier<VillagerType> register(String key) {